backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
databases[postgresql]==0.9.0
pydantic==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
aioredis==2.0.1
email-validator==2.2.0